# OrigenX/BackEnd/openapi.yaml
openapi: 3.0.3
info:
  title: The Artisanal Connection API
  description: |
    API REST para The Artisanal Connection. Conecta productores de cafe colombiano con consumidores finales.

    ## Autenticacion
    La mayoria de endpoints requieren un Firebase ID Token en el header:
    `Authorization: Bearer <firebase_id_token>`

    ## Roles
    - **CONSUMER**: consumidor final que navega el catalogo y realiza pedidos.
    - **PRODUCER**: productor de cafe que gestiona productos y ventas.
    - **ADMIN**: administrador de la plataforma.

    ## Formato de errores
    Todos los errores siguen el formato uniforme:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Descripcion del error",
        "fields": [{"field": "email", "message": "El correo ya esta registrado"}]
      }
    }
    ```
  version: 0.1.0
  contact:
    name: The Artisanal Connection
servers:
  - url: http://localhost:8000
    description: Servidor de desarrollo local
  - url: https://api.marketplacecafetero.com
    description: Servidor de produccion
...
security:
  - BearerAuth: []
paths:
...
  /api/producers/{producer_id}:
    get:
      tags: [producers]
      summary: Obtener perfil de productor
      description: Retorna el perfil publico de un productor. Accesible por cualquier usuario. Requerimientos 4.1, 4.4.
      operationId: getProducerProfile
      security: []
      parameters:
        - $ref: "#/components/parameters/ProducerId"
      responses:
        "200":
          description: Perfil del productor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ProducerProfileData"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [producers]
      summary: Actualizar perfil de productor
      description: Crea o actualiza el perfil de un productor. Solo el Productor dueno puede actualizar su propio perfil. Retorna HTTP 400 si farmName esta vacio. Retorna HTTP 403 si el usuario autenticado no es el dueno del perfil. Requerimientos 4.1-4.4.
      operationId: updateProducerProfile
      parameters:
        - $ref: "#/components/parameters/ProducerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProducerProfileCreate"
            example:
              farm_name: Finca La Esperanza
              region: Huila
              description: Productores de cafe especial desde 1985.
              contact_info: contacto@fincaesperanza.com
      responses:
        "200":
          description: Perfil actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Perfil actualizado exitosamente.
                  data:
                    $ref: "#/components/schemas/ProducerProfileData"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID Token obtenido mediante Firebase Client SDK. Incluir en el header Authorization: Bearer <token>"
  parameters:
...
    ProducerId:
      name: producer_id
      in: path
      required: true
      description: Identificador unico del productor (Firebase UID)
      schema:
        type: string
        example: uid789ghi
...
  schemas:
    ErrorField:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: El correo ya esta registrado
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Descripcion del error
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ErrorField"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
...
    ProducerProfileCreate:
      type: object
      required: [farm_name]
      properties:
        farm_name:
          type: string
          minLength: 1
          description: Nombre de la finca (obligatorio)
          example: Finca La Esperanza
        region:
          type: string
          nullable: true
          example: Huila
        description:
          type: string
          nullable: true
          example: Productores de cafe especial desde 1985.
        contact_info:
          type: string
          nullable: true
          example: contacto@fincaesperanza.com
...
    ProducerProfileData:
      type: object
      description: Datos del perfil de un productor
      properties:
        user_id:
          type: string
          example: uid789ghi
        farm_name:
          type: string
          example: Finca La Esperanza
        region:
          type: string
          nullable: true
          example: Huila
        description:
          type: string
          nullable: true
          example: Productores de cafe especial desde 1985.
        contact_info:
          type: string
          nullable: true
          example: contacto@fincaesperanza.com
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
...
  responses:
    ValidationError:
      description: Datos de entrada invalidos (HTTP 400)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Los datos de entrada son invalidos.
              fields:
                - field: email
                  message: El correo electronico no tiene un formato valido.
                - field: password
                  message: La contrasena debe tener al menos 8 caracteres.
    UnauthorizedError:
      description: No autenticado - JWT ausente o invalido (HTTP 401)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Token de autenticacion ausente o invalido.
    ForbiddenError:
      description: No autorizado - rol insuficiente o recurso ajeno (HTTP 403)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: No tienes permiso para realizar esta accion.
    NotFoundError:
      description: Recurso no encontrado (HTTP 404)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: El recurso solicitado no fue encontrado.
...
    InternalError:
      description: Error interno del servidor (HTTP 500)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: Error interno del servidor. Por favor intenta de nuevo.
...